Start export file creation

Has server-side side-effect of starting an export job for the target questionnaire. The function parameters provide a way to describe which interviews should be exported and for which questionnaire.

Usage

start_export(
  qnr_id,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)

Arguments

qnr_id

Character. Questionnaire ID. Format: qnr_id$version

export_type

Character. Type of data to export. Values: c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")

interview_status

Character. Status of interviews to export. Values: c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")

from

Character. Start date for interviews to export. date-time, UTC

to

Character. End date for interviews to export. date-time, UTC

access_token

Access token to external storage. Relevant only if external cloud storage is destination for the export file.

refresh_token

Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.

storage_type

Character. External storage type, if relevant. Values: c("Dropbox", "OneDrive", "GoogleDrive")

translation_id

Character. Translation ID for variable and value labels to include in export files.

include_meta

Logical. If TRUE, include questionnaire metadata in export file.

server

Character. Full server web address (e.g., https://demo.mysurvey.solutions, https://my.domain)

workspace

Character. Name of the workspace whose export jobs to start. In workspace list, value of NAME, not DISPLAY NAME, for the target workspace.

user

Character. API user name

password

Character. API password

Value

Job ID of export process started.

Details

Wrapper for POST /api/v2/export endpoint