Start export file creation

Has server-side side-effect of starting an export job for the target questionnaire. The function parameters provide a way to describe which interviews should be exported and for which questionnaire.

start_export(
  qnr_id,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)

Arguments

qnr_id
Questionnaire ID. Format: qnr_id$version

export_type
Type of data to export. Values: c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")

interview_status
Status of interviews to export. Values: c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")

from
Start date for interviews to export. date-time, UTC

to
End date for interviews to export. date-time, UTC

access_token
Access token to external storage. Relevant only if external cloud storage is destination for the export file.

refresh_token
Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.

storage_type
External storage type, if relevant. Values: c("Dropbox", "OneDrive", "GoogleDrive")

translation_id
Translation ID for variable and value labels to include in export files.

include_meta
Logical. If TRUE, include questionnaire metadata in export file.

server
Full server web address (e.g., https://demo.mysurvey.solutions, https://my.domain)

user
API user name

password
API password

Value

Job ID of export process started.

Details

Wrapper for POST /api/v2/export endpoint